Serial correlation between grab samples

This function calculates the resulting serial correlation between grab samples each having r primary increments with original serial correlation d.

correlation_grab(r, p, d)

Arguments

r
number of primary increments in a grab sample or grab sample size

p
limiting fraction or proportion of contaminated increments

d
serial correlation of contamination between the primary increments

Value

Serial correlation between grab samples

Details

The serial correlation between blocks (grab samples) is given by \(d_g\) as $$d_g = [dp(1-p(1-d))^{r-1}]/p_d$$ where \(p_d\) is the probability of detection in any of the block (grab sample) which is calculated by using prob_detect_single_grab.

See also

Examples

r <-  25
p <-  0.005
d <-  0.99
correlation_grab(r, p, d)

#> [1] 0.7982905