Probability of detection for grab sample selection

This package provides the probability of detection calculation for grab samples selection by various method of samplings such as systematic or random and also it is useful to generate the comparison curves. Moreover, this package calculates the probability of acceptance calculations based on suitable microbiological distributions such as Poisson gamma or Lognormal or Poisson lognormal and also provides a comparison based on OC curves with different sampling schemes. Most of the researchers have studied the uncorrelated case, but in this study, we have a high spatial correlation between contamination of primary increments. For this package development, we used default standard deviation as 0.8 and also spatial correlation not affected for composite mean for the probability of acceptance calculation. A future version will be included deeply study about variability effects in grab sample selection.

References

Bhat, U., & Lal, R. (1988). Number of successes in Markov trials. Advances in Applied Probability, 20(3), 677-680.

Jongenburger, I., Besten, H.M., & Zwietering, M.H. (2015). Statistical aspects of food safety sampling. Annual review of food science and technology, 6, 479-503.

Mussida, A., Vose, D. & Butler, F. Efficiency of the sampling plan for Cronobacter spp. assuming a Poisson lognormal distribution of the bacteria in powder infant formula and the implications of assuming a fixed within and between-lot variability, Food Control, Elsevier, 2013 , 33 , 174-185.

Van Schothorst, M., Zwietering, M., Ross, T., Buchanan, R. & Cole, M., Relating microbiological criteria to food safety objectives and performance objectives Food Control , 2009 , 20 , 967-979.

Author

Mayooran Thevaraja, Kondaswamy Govindaraju, Mark Bebbington